(PE-35979) resolve logging issues
This updates the version of logback to the latest.  This is necessary for the jetty integration
to work correctly. It also reenables the MDC log converter and handler and injects them into
the requests.

One test was updated to include more bytes in the response.  This seems to be new things that
Jetty 10 is including that Jetty 9 did not, and didn't appear to be harmful.