(PE-35979)  fix proxy test issues
This updates the proxy handling to work correctly with jetty 10.

The `SslContextFactory$Server` in jetty 9 could be used for both server and
client configuration.  In Jetty 10, there is a separate class specifically for
client configuraton that must be used.  This updates the interfaces to configure
both a server and client factory at setup time to allow the client to correctly
use the ssl-configuraton with the requests that it is proxying.

The `failure-callback` behavior isn't simple nor straightforward as
failure modes in jetty have changed.  For example, the tests that
were checking the failure-callback are now generating an "IllegalArgumentException"
that prevents the normal error handling routines from being used.
Since this functionality isn't being used in any Puppet products,
it is being removed.