(PE-35987) Fix logging tests with SSLv3 warning
This fixes the issues with the tests with the SSLV3 warning by removing
the warning from the client initialization.  The "logged" test macro
only allows one match for the log entry in the results, and the client
is producing the same entry as the server.  The client warning was removed as
the server is the primary concern.