(PE-37252) discontinue use of logback-access
This disables the use of logback-access because logback-access is not
compatible with Jetty 10.  See https://github.com/jetty/jetty.project/issues/5996

This changes the implementation to use a CustomRequestLog in jetty with the SLF4JLogWriter
until a better solution can be determined.  This means that logging won't obey the
logback-access configuration specified.