(PE-37256) restore access-logging
Previously access-logging was switched to the Jetty CustomLogger, which
wrote to the main system logs by default, and didn't provide any MDC
logging.

This changes the behavior to reenable access-logging by using the logback `RequestLogImpl` directly
in the Jetty Servers `requestLog` setting.  This bypasses the previous issues with the data
being settled, and allows the configuration and MDC logging to work correctly.

Unfortunately, `RequestLogImpl` can't be used directly, so this provides an alternative
implementation that works around some compile time compatibility with logback-access and Jetty 10
by providing modified versions of the JettyModernServerAdapter and RequestLogImpl